+++ src/views/MainPage.vue
@@ -30,66 +30,20 @@
   </div>
   <div class="divider" />
   <div class="travel-course">
-    <h2 class="mb-3">여행 후기 게시판</h2>
-    <div class="latestReview-cards">
-      <div
-        v-for="(review, index) in latestReviews"
-        :key="index"
-        class="latestReview-card"
-      >
-        <div class="latestReview-image">
-          <img
-            :src="
-              review.reviewImages.length > 0
-                ? review.reviewImages[0].imageUri
-                : 'default-image-url'
-            "
-            alt="ReviewImage"
-          />
-        </div>
-        <h5>{{ review.reviewTitle }}</h5>
-        <div class="latestReview-content" v-html="review.reviewContent"></div>
-      </div>
-      <router-link class="more-reviews" to="/reviews">
-        여행 후기 더보기</router-link
-      >
-    </div>
+    <LatestReviews />
   </div>
   <div class="divider" />
 </template>
 
 <script>
+import LatestReviews from "../components/Review/LatestReviews.vue";
 // @ is an alias to /src
 /* eslint-disable */
 export default {
   name: "MainPage",
-  components: {},
-  data() {
-    return {
-      allReviews: [],
-      latestReviews: [],
-    };
-  },
-  created() {
-    this.fetchAllReviews();
-  },
-  methods: {
-    async fetchAllReviews() {
-      try {
-        const response = await this.$axios.get("/api/reviews");
-        this.allReviews = response.data.content;
-        this.latestReviews = this.allReviews.slice(0, 4); //전체리뷰에서 최신 3개 선택
-      } catch (error) {
-        if (error.response && error.response.data) {
-          //백엔드에서 보낸 에러메시지 표시
-          alert(`${error.response.data.message}`);
-          console.error(error.response.data.message);
-        } else {
-          console.error("에러났어요 : " + error);
-        }
-      }
-    },
-  },
+  components: { LatestReviews },
+
+  methods: {},
 };
 </script>
 <style scoped>
@@ -134,44 +88,5 @@ export default {
   width: 100%; /* 카드의 너비를 화면 전체로 설정 */
   display: flex;
   justify-content: center; /* 내부 요소를 가로축 중앙에 배치 */
-}
-.latestReview-cards {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px; /* 카드 사이의 간격 */
-}
-
-.latestReview-card {
-  flex: 1;
-  min-width: 15%; /* 최소 너비 */
-  background-color: #fff;
-  border-radius: 10px; /* 모서리 둥글게 */
-  overflow: hidden; /* 내용이 넘칠 경우 숨김 */
-  margin: 10px;
-  padding: 20px;
-}
-.latestReview-image img {
-  cursor: pointer;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  margin-bottom: 30px;
-}
-
-.latestReview-content {
-  display: -webkit-box;
-  -webkit-line-clamp: 3;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.more-reviews {
-  display: block;
-  text-align: center;
-  justify-content: center;
-  margin-top: 20px;
-  text-decoration: none;
-  /* 더보기 링크 스타일 */
 }
 </style>
